refactor: splits documentatiegenerator op in losse klassen
